requirements.txt
streamlit==1.29.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
pandas==2.1.4
plotly==5.17.0
openpyxl==3.1.2
python-dateutil==2.8.2
regex==2023.10.3
opencv-python==4.8.1.78